move character_inventory in chracter module
