+++ world_server/src/character/character_database.rs
@@ -1,4 +1,4 @@
-use crate::data::{DataStorage, TutorialFlags, WorldZoneLocation, INVENTORY_SLOT_BAG_0};
+use crate::data::{DataStorage, TutorialFlags, WorldZoneLocation};
 use crate::packet::ServerMessageExt;
 use crate::prelude::*;
 use crate::world::prelude::*;
@@ -7,6 +7,8 @@ use std::time::{SystemTime, UNIX_EPOCH};
 use wow_dbc::Indexable;
 use wow_world_messages::wrath::{Area, Class, Gender, Map, MovementInfo, Power, Race, Vector3d, SkillInfoIndex,SkillInfo, SMSG_INITIAL_SPELLS, InitialSpell, SMSG_UPDATE_OBJECT, Object, Object_UpdateType, UpdateItemBuilder, UpdateMask, MovementBlock, MovementBlock_UpdateFlag, UpdatePlayerBuilder, SMSG_ITEM_PUSH_RESULT, VisibleItemIndex, VisibleItem};
 use wow_world_base::wrath::{RaceClass, ObjectType, ItemSlot, NewItemChatAlert, NewItemCreationType, NewItemSource};
+
+use super::character_inventory::INVENTORY_SLOT_BAG_0;
 impl super::Character {
     pub(super) async fn load_from_database_internal(&mut self, world: &World, data_storage: &DataStorage) -> Result<()> {
         let character_id = self.get_guid().guid() as u32;
